Pierwsza część dużego update'a. WAŻNE: proszę wyczyścić katalog z buildem z katalogów typu "core", "corelib", "utils" i "plugins". Pominięcie tego kroku może spowodować, że zostaną tam nagłówki konfiguracyjne i podczas budowania będą one przesłaniać te właściwe. 1) Zmiany nazw zmiennych CMake'a:   CONFIG_CORELIB_LOG4CXX_ENABLED -> CONFIG_CORE_ENABLE_LOG4CXX (domyślnie off)   UTILS_ENABLE_LEAK_DETECTION (albo podobnie) -> CONFIG_CORE_ENABLE_LEAK_DETECTION (domyślnie off) 2) Dodanie możliwości dodawania plików zasobów Qt (niedługo będzie udokumentowane) 3) Usprawnienia widgetu konsoli (m.in. dodanie ikon w katalogu resources); konsola działa nawet przy wyłączonej fladze CONFIG_CORE_ENABLE_LOG4CXX 4) Przekierowywanie logowania Qt do naszego systemu 5) Usunięcie (tak!) biblioteki utils oraz znacznej części pluginu video; obie te części (ta druga od teraz nazywana vidlib) znajdą się w osobnym projekcie EDR Utils, w który będzie skupiony na bibliotekach oraz na wyczerpującym ich testowaniu; skutkiem "ubocznym" trzeba wskazać teraz lokalizację obu bibliotek (UTILS_* oraz VIDLIB_*) tak jak dla wszystkich zewnętrznych zależności; mojego builda wrzucę niebawem na assemblę do sekcji Files, niedługo Michał doda odpowiednią procedurę do CI
git-svn-id: https://localhost:8443/svn/hm-mde-1@273 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/resources/shaders/textureRect_yuv_to_rgb.frag
@@ -9,13 +9,13 @@ const float yScale = 2.0/3.0;
 const float uvScale = 1.0-yScale;
 
 // rozmiar tekstury (pobieranie instrukcjami texture*Size nie zawsze dzia³a)
-uniform ivec2 texture_size;
+uniform vec2 texture_size;
 
 void main(void)
 {
 	// rozmiar tekstury
-	int texWidth = texture_size.x;
-	int texHeight = texture_size.y;
+	float texWidth = texture_size.x;
+	float texHeight = texture_size.y;
 
 	// wyznaczenie koordynatów
 	vec2 stY = gl_TexCoord[0].st;
